+++ docs/changes.rst
@@ -9,6 +9,11 @@ skops Changelog
     :depth: 1
     :local:
 
+v0.3
+----
+- Utility function to add arbitrary files to be uploaded to the hub by using
+  :func:`.hub_utils.add_files`. :pr:`123` by `Benjamin Bossan`_.
+
 v0.2
 ----
 - Tables, e.g. cross-validation results, can now be added to model cards using